DestroysEarnedChallenges action implemented
