+++ app/services/cancels_course_membership.rb
@@ -1,6 +1,7 @@
 require "light-service"
 require_relative "cancels_course_membership/destroys_assignment_weights"
 require_relative "cancels_course_membership/destroys_earned_badges"
+require_relative "cancels_course_membership/destroys_earned_challenges"
 require_relative "cancels_course_membership/destroys_grades"
 require_relative "cancels_course_membership/destroys_membership"
 require_relative "cancels_course_membership/destroys_submissions"
@@ -60,20 +61,6 @@ class CancelsCourseMembership
 
   def self.removes_group_memberships(membership)
     GroupMembership.for_course(membership.course)
-      .for_student(membership.user)
-      .destroy_all
-    self
-  end
-
-  def self.removes_predicted_earned_challenges(membership)
-    PredictedEarnedChallenge.for_course(membership.course)
-      .for_student(membership.user)
-      .destroy_all
-    self
-  end
-
-  def self.removes_submissions(membership)
-    Submission.for_course(membership.course)
       .for_student(membership.user)
       .destroy_all
     self
